LayoutTests/platform/mac/plugins/plugin-clip-subframe-expected.txt

CONSOLE MESSAGE: line 3: NPP_SetWindow: NULL window, Rect {18, 218, 300, 150}, Clip Rect {0, 600, 0, 600}, Type 0
CONSOLE MESSAGE: line 3: NPP_SetWindow: NULL window, Rect {18, 218, 300, 150}, Clip Rect {18, 218, 108, 368}, Type 0
